Moved trailer dumping to read
