+++ Makefile
@@ -1,19 +1,3 @@
-all: read read_trailer
+all:
+	g++ -O2 -Wall -o read read.cpp
 
-read: read.o common.o
-	g++ -O2 -Wall -o read read.o common.o
-
-read.o: read.cpp
-	g++ -O2 -Wall -c -o read.o read.cpp
-
-read_trailer: read_trailer.o common.o
-	g++ -O2 -Wall -o read_trailer read_trailer.o common.o
-
-read_trailer.o: read_trailer.cpp
-	g++ -O2 -Wall -c -o read_trailer.o read_trailer.cpp
-
-common.o: common.cpp
-	g++ -O2 -Wall -c -o common.o common.cpp
-
-clean:
-	rm -f read read_trailer *.o
